material update

+++ TitanCore/src/VertexIndexBufferData.cpp
@@ -10,30 +10,35 @@ namespace Titan
 		mMgr = mgr ? mgr : HardwareBufferManager::getSingletonPtr();
 		vertexDecl = mMgr->createVertexDeclaration();
 		vertexBufferBinding = mMgr->createVertexBufferBinding();
+		mDeleteDclBinding = true;
 	}
-	//-------------------------------------------------------------//
+	//-------------------------------------------------------------------------------//
 	VertexData::VertexData(VertexDeclaration* decl, VertexBufferBinding* binding)
 		: vertexStart(0), vertexCount(0)
 	{
 		mMgr = HardwareBufferManager::getSingletonPtr();
 		vertexDecl = decl;
 		vertexBufferBinding = binding;
+		mDeleteDclBinding = false;
 	}
-	//-------------------------------------------------------------//
+	//-------------------------------------------------------------------------------//
 	VertexData::~VertexData()
 	{
-		mMgr->destroyVertexBufferBinding(vertexBufferBinding);
-		mMgr->destroyVertexDeclaration(vertexDecl);
+		if (mDeleteDclBinding)
+		{
+			mMgr->destroyVertexBufferBinding(vertexBufferBinding);
+			mMgr->destroyVertexDeclaration(vertexDecl);
+		}
 	}
 	
 
 
-	//-------------------------------------------------------------//
+	//-------------------------------------------------------------------------------//
 	IndexData::IndexData()
 		: indexStart(0), indexCount(0)
 	{
 	}
-	//-------------------------------------------------------------//
+	//-------------------------------------------------------------------------------//
 	IndexData::~IndexData()
 	{
 	}
